+++ dcmrt/libsrc/drtdss.cc
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2014, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Source file for class DRTDigitalSignaturesSequence
  *
- *  Generated automatically from DICOM PS 3.3-2017e
- *  File created on 2017-12-05 09:30:54
+ *  Generated automatically from DICOM PS 3.3-2007
+ *  File created on 2014-03-15 16:58:36
  *
  */
 
@@ -446,7 +446,7 @@ OFBool DRTDigitalSignaturesSequence::isValid() const
 }
 
 
-size_t DRTDigitalSignaturesSequence::getNumberOfItems() const
+unsigned long DRTDigitalSignaturesSequence::getNumberOfItems() const
 {
     return SequenceOfItems.size();
 }
@@ -476,12 +476,12 @@ OFCondition DRTDigitalSignaturesSequence::gotoNextItem()
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::gotoItem(const size_t num, OFListIterator(Item *) &iterator)
+OFCondition DRTDigitalSignaturesSequence::gotoItem(const unsigned long num, OFListIterator(Item *) &iterator)
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -496,12 +496,12 @@ OFCondition DRTDigitalSignaturesSequence::gotoItem(const size_t num, OFListItera
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::gotoItem(const size_t num, OFListConstIterator(Item *) &iterator) const
+OFCondition DRTDigitalSignaturesSequence::gotoItem(const unsigned long num, OFListConstIterator(Item *) &iterator) const
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -516,7 +516,7 @@ OFCondition DRTDigitalSignaturesSequence::gotoItem(const size_t num, OFListConst
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::gotoItem(const size_t num)
+OFCondition DRTDigitalSignaturesSequence::gotoItem(const unsigned long num)
 {
     return gotoItem(num, CurrentItem);
 }
@@ -552,7 +552,7 @@ const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getCurre
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::getItem(const size_t num, Item *&item)
+OFCondition DRTDigitalSignaturesSequence::getItem(const unsigned long num, Item *&item)
 {
     OFListIterator(Item *) iterator;
     OFCondition result = gotoItem(num, iterator);
@@ -562,7 +562,7 @@ OFCondition DRTDigitalSignaturesSequence::getItem(const size_t num, Item *&item)
 }
 
 
-DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(const size_t num)
+DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(const unsigned long num)
 {
     OFListIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -572,7 +572,7 @@ DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(const 
 }
 
 
-const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(const size_t num) const
+const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(const unsigned long num) const
 {
     OFListConstIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -582,13 +582,13 @@ const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::getItem(
 }
 
 
-DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::operator[](const size_t num)
+DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::operator[](const unsigned long num)
 {
     return getItem(num);
 }
 
 
-const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::operator[](const size_t num) const
+const DRTDigitalSignaturesSequence::Item &DRTDigitalSignaturesSequence::operator[](const unsigned long num) const
 {
     return getItem(num);
 }
@@ -611,7 +611,7 @@ OFCondition DRTDigitalSignaturesSequence::addItem(Item *&item)
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::insertItem(const size_t pos, Item *&item)
+OFCondition DRTDigitalSignaturesSequence::insertItem(const unsigned long pos, Item *&item)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
@@ -634,7 +634,7 @@ OFCondition DRTDigitalSignaturesSequence::insertItem(const size_t pos, Item *&it
 }
 
 
-OFCondition DRTDigitalSignaturesSequence::removeItem(const size_t pos)
+OFCondition DRTDigitalSignaturesSequence::removeItem(const unsigned long pos)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
