+++ dcmrt/libsrc/drtdvrrs.cc
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2014, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Source file for class DRTDVHReferencedROISequence
  *
- *  Generated automatically from DICOM PS 3.3-2017e
- *  File created on 2017-12-05 09:30:54
+ *  Generated automatically from DICOM PS 3.3-2007
+ *  File created on 2014-03-15 16:58:36
  *
  */
 
@@ -260,7 +260,7 @@ OFBool DRTDVHReferencedROISequence::isValid() const
 }
 
 
-size_t DRTDVHReferencedROISequence::getNumberOfItems() const
+unsigned long DRTDVHReferencedROISequence::getNumberOfItems() const
 {
     return SequenceOfItems.size();
 }
@@ -290,12 +290,12 @@ OFCondition DRTDVHReferencedROISequence::gotoNextItem()
 }
 
 
-OFCondition DRTDVHReferencedROISequence::gotoItem(const size_t num, OFListIterator(Item *) &iterator)
+OFCondition DRTDVHReferencedROISequence::gotoItem(const unsigned long num, OFListIterator(Item *) &iterator)
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -310,12 +310,12 @@ OFCondition DRTDVHReferencedROISequence::gotoItem(const size_t num, OFListIterat
 }
 
 
-OFCondition DRTDVHReferencedROISequence::gotoItem(const size_t num, OFListConstIterator(Item *) &iterator) const
+OFCondition DRTDVHReferencedROISequence::gotoItem(const unsigned long num, OFListConstIterator(Item *) &iterator) const
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        size_t idx = num + 1;
+        unsigned long idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -330,7 +330,7 @@ OFCondition DRTDVHReferencedROISequence::gotoItem(const size_t num, OFListConstI
 }
 
 
-OFCondition DRTDVHReferencedROISequence::gotoItem(const size_t num)
+OFCondition DRTDVHReferencedROISequence::gotoItem(const unsigned long num)
 {
     return gotoItem(num, CurrentItem);
 }
@@ -366,7 +366,7 @@ const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getCurrent
 }
 
 
-OFCondition DRTDVHReferencedROISequence::getItem(const size_t num, Item *&item)
+OFCondition DRTDVHReferencedROISequence::getItem(const unsigned long num, Item *&item)
 {
     OFListIterator(Item *) iterator;
     OFCondition result = gotoItem(num, iterator);
@@ -376,7 +376,7 @@ OFCondition DRTDVHReferencedROISequence::getItem(const size_t num, Item *&item)
 }
 
 
-DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(const size_t num)
+DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(const unsigned long num)
 {
     OFListIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -386,7 +386,7 @@ DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(const si
 }
 
 
-const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(const size_t num) const
+const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(const unsigned long num) const
 {
     OFListConstIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -396,13 +396,13 @@ const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::getItem(co
 }
 
 
-DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::operator[](const size_t num)
+DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::operator[](const unsigned long num)
 {
     return getItem(num);
 }
 
 
-const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::operator[](const size_t num) const
+const DRTDVHReferencedROISequence::Item &DRTDVHReferencedROISequence::operator[](const unsigned long num) const
 {
     return getItem(num);
 }
@@ -425,7 +425,7 @@ OFCondition DRTDVHReferencedROISequence::addItem(Item *&item)
 }
 
 
-OFCondition DRTDVHReferencedROISequence::insertItem(const size_t pos, Item *&item)
+OFCondition DRTDVHReferencedROISequence::insertItem(const unsigned long pos, Item *&item)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
@@ -448,7 +448,7 @@ OFCondition DRTDVHReferencedROISequence::insertItem(const size_t pos, Item *&ite
 }
 
 
-OFCondition DRTDVHReferencedROISequence::removeItem(const size_t pos)
+OFCondition DRTDVHReferencedROISequence::removeItem(const unsigned long pos)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
